pyproject.toml
[tool.poetry]
name = "lidarmeasurementdata"
version = "0.1.0"
description = ""
authors = ["ICheered <tjeerd992@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
rosbags = "^0.9.13"
bagpy = "^0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
